changed coexpression calc approach to einsum based

+++ src/coexpression/spearman.py
@@ -12,6 +12,7 @@ import concurrent.futures as cf
 import multiprocessing as mp
 from coexpression import ensemble
 import scipy
+from einsumt import einsumt
 
 def precalc(expmat_path, Tid2Gid_dict, k_cluster_assignment_dict, k, delimiter="\t", workers=2):
     assignment= k_cluster_assignment_dict[k]
@@ -46,32 +47,39 @@ def precalc(expmat_path, Tid2Gid_dict, k_cluster_assignment_dict, k, delimiter="
 
 
 
-def calc_job_k(source_array, target_array, shared_nominators_dict, shared_denominators_dict, cluster):
-    cor_values = np.sum(shared_nominators_dict[cluster][source_array] * shared_nominators_dict[cluster][target_array], axis=1)/(shared_denominators_dict[cluster][source_array]* shared_denominators_dict[cluster][target_array])
+def calc_job_k(source_array, target_array, shared_nominators_dict, shared_denominators_dict, cluster, threads):
+    #cor_values = np.sum(shared_nominators_dict[cluster][source_array] * shared_nominators_dict[cluster][target_array], axis=1)/(shared_denominators_dict[cluster][source_array]* shared_denominators_dict[cluster][target_array])
+    numerator = einsumt('ij,ij->i', np.take(shared_nominators_dict[cluster], source_array , axis = 0) , np.take(shared_nominators_dict[cluster], target_array , axis = 0), pool =threads)
+    denominator = np.einsum('i,i->i',  np.take(shared_denominators_dict[cluster], source_array), np.take(shared_denominators_dict[cluster], target_array))
+    cor_values = numerator / denominator
     return cor_values
 
 
-def calc_targeted(k, path, edges , genes, gene_dict, nominators_dict, denominators_dict , workers = 2):
-    manager = mp.Manager()
-    shared_nominators_dict = manager.dict(nominators_dict)
-    shared_denominators_dict = manager.dict(denominators_dict)
+def calc_targeted(k, path, edges , genes, gene_dict, nominators_dict, denominators_dict , threads, workers = 2):
     #add header if creating new file
-    with open(path, "w") as f:
-        f.write("Edge\tcluster_cor\tMax,Avg,RAvg,RWA,RRWA\n")
+    if not os.path.exists(path):
+        with open(path, "w") as f:
+            f.write("Edge\tcluster_cor\tMax,Avg,RAvg,RWA,RRWA\n")
     source_array , target_array= [], []
 
     for edge in list(edges):
         source, target = edge.split("--")
         source_array.append(gene_dict[source])
         target_array.append(gene_dict[target])
+    source_array = np.array(source_array)
+    target_array = np.array(target_array)
 
     ALL_cor_means = []
-    with cf.ProcessPoolExecutor(max_workers=workers) as executor:
-        results = [executor.submit(calc_job_k, source_array, target_array, shared_nominators_dict, shared_denominators_dict, cluster) for cluster in range(k)]
-        #results = [executor.submit(calc_job_k, source_array, target_array, nominators_dict[cluster], denominators_dict[cluster], cluster) for cluster in range(k)]
-        for f in cf.as_completed(results):
-            cor_means = f.result()
-            ALL_cor_means.append(cor_means)
+    
+    for cluster in range(k):
+        cor_means = calc_job_k( source_array, target_array, nominators_dict, denominators_dict, cluster, threads)
+        ALL_cor_means.append(cor_means)
+    
+    #with cf.ProcessPoolExecutor(max_workers=workers) as executor:
+    #    results = [executor.submit(calc_job_k, source_array, target_array, nominators_dict, denominators_dict, cluster, threads) for cluster in range(k)]
+    #    for f in cf.as_completed(results):
+    #        cor_means = f.result()
+    #        ALL_cor_means.append(cor_means)
     ALL_cor_means = np.array(ALL_cor_means)
     batch_ensemble_scores = []
     for mode in ["Max", "Avg", "RAvg", "RWA", "RRWA"]:
@@ -89,11 +97,16 @@ def calc_targeted(k, path, edges , genes, gene_dict, nominators_dict, denominato
 def calc_untargeted():
     pass
         
-def optimize_k(k, positive_met_edges_cor_path, negative_met_edges_cor_path ,expmat_path, Tid2Gid_dict,  k_cluster_assignment_dict, delim, workers, positive_met_edges, negative_met_edges_unpacked):
-       genes, gene_dict, nominators_dict, denominators_dict = precalc(expmat_path, Tid2Gid_dict, k_cluster_assignment_dict, k, delimiter=delim, workers=workers)
+def optimize_k(k, positive_met_edges_cor_path, negative_met_edges_cor_path ,expmat_path, Tid2Gid_dict,  k_cluster_assignment_dict, delim, workers, positive_met_edges, negative_met_edges_unpacked, threads):
+    genes, gene_dict, nominators_dict, denominators_dict = precalc(expmat_path, Tid2Gid_dict, k_cluster_assignment_dict, k, delimiter=delim, workers=workers)
+    manager = mp.Manager()
+    shared_nominators_dict = manager.dict(nominators_dict)
+    shared_denominators_dict = manager.dict(denominators_dict)
+    print("Calculating and writing correlations of positive edges...")
 
-       print("Calculating and writing correlations of positive edges...")
-
-       calc_targeted(k, positive_met_edges_cor_path, positive_met_edges , genes, gene_dict, nominators_dict, denominators_dict ,workers = workers)
-       print("Calculating and writing correlations of negative edges...")
-       calc_targeted(k, negative_met_edges_cor_path, negative_met_edges_unpacked , genes, gene_dict, nominators_dict,denominators_dict , workers = workers)
+    calc_targeted(k, positive_met_edges_cor_path, positive_met_edges , genes, gene_dict, shared_nominators_dict, shared_denominators_dict, threads ,workers = workers)
+    
+    print("Calculating and writing correlations of negative edges...")
+    negative_met_edges_chunks = [negative_met_edges_unpacked[x: x+ 200000 ] for x in range(0, len(negative_met_edges_unpacked), 200000)]
+    for chunk in negative_met_edges_chunks:
+        calc_targeted(k, negative_met_edges_cor_path, chunk , genes, gene_dict, shared_nominators_dict, shared_denominators_dict , threads, workers = workers)
